combined location hooks into one
