+++ src/index.tsx
@@ -19,16 +19,47 @@ const { Header } = Layout;
 
 import "./style.css";
 import { setIsPlaying } from "./state/selection/actions";
+import { clearSimulariumFile } from "./state/metadata/actions";
+import {
+    getSimulariumController,
+    getSimulariumFile,
+} from "./state/metadata/selectors";
 
 export const store = createReduxStore();
-
+interface LocationWithState extends Location {
+    state: {
+        localFile?: boolean;
+    };
+}
 function useLocationChangeToStopPlay() {
-    const location = useLocation();
+    const location = useLocation() as LocationWithState;
     const dispatch = useDispatch();
 
     React.useEffect(() => {
-        // if we're navigating away from the viewer, stop playing
-        if (location.pathname !== VIEWER_PATHNAME) {
+        if (location.pathname === VIEWER_PATHNAME) {
+            /**
+             * Gets called every time the app navigates to the Simularium page.
+             * There are 2 possible url types and 3 different states the viewer should be in
+             * when we get there:
+             * 1. Empty viewer, by clicking "launch viewer" or "load your own data" card; url: '/viewer'
+             * 2. Loading network file; url: '/viewer?trajFileName=FILE_NAME'
+             * 3. Loading a local file through the dropdown, will be reflected in the location state; url: '/viewer'
+             */
+            const state = store.getState();
+            const controller = getSimulariumController(state);
+            const simFile = getSimulariumFile(state);
+            // got here from the "load local file button" so the app is going to
+            // `/viewer`, but the loadFile action will take care of resetting state
+            // if the user clicks "Open"
+            if (location.state && location.state.localFile) {
+                return;
+            }
+            if (!location.search && controller && simFile.name) {
+                // going to /viewer, clear out any existing files
+                dispatch(clearSimulariumFile({ newFile: false }));
+            }
+        } else {
+            // if we're navigating away from the viewer, stop playing
             dispatch(setIsPlaying(false));
         }
     }, [location]);
